refactor: test API
